add(convert): convert_to_double() use ft_strtod(); (separate file)
